hackathon/requirements.txt
fastapi
uvicorn
pytesseract
pdf2imag
pinecone
llmware
sklearn
ollama